added opirations

+++ libft/stack/stack.h
@@ -10,6 +10,6 @@ typedef struct s_stack
 t_stack	*ft_stnew(long n);
 void	ft_stpush(t_stack **stack, t_stack *content);
 void	ft_stpop(t_stack **stack);
-int 	ft_stpeek(t_stack *stack);
+long 	ft_stpeek(t_stack *stack);
 int		ft_stisempty(t_stack *stack);
 int		ft_stsize(t_stack *stack);
